Agent的概念、原理与构建模式/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.2.1",
    "openai>=1.91.0",
    "python-dotenv>=1.1.1",
    "ollama>=0.5.1",
    "google-genai>=1.27.0",
    "langchain-google-genai>=0.1.0",
    "langchain-ollama>=0.1.0",
]
[tool.uv.workspace]
members = [
    "snake3",
]
